Use single-precision math functions throughout (sinf, cosf, sqrtf, etc.)
ARM11 VFPv2 only has single-precision hardware — double-precision
operations fall back to software emulation at 2-4x the cost. The
codebase used double-precision math functions (sin, cos, sqrt, fabs,
floor, ceil, pow, asin, acos, atan2, tan) everywhere despite all
values being float. Replaced 307 calls across 35 files with their
float-suffixed C99 equivalents.

Also updated SIND/COSD/SINR/COSR/ATAND/FMOD/FDIV macros in new3d.h.
Gamma table in render_gl_shared.c intentionally kept as double pow().

+++ ailib.c
@@ -246,7 +246,7 @@ void AI_AimAtTarget( MATRIX * InvMat , VECTOR * SPos, VECTOR * TPos )
 	ApplyMatrix( InvMat, &WantedDir, &TempDir );			/* Calc Direction Vector */
 
 	NormaliseVector( &TempDir );
-	Angle = (float) acos( TempDir.x );
+	Angle = (float) acosf( TempDir.x );
 	Angle = 90.0F - R2D( Angle );
 	if( TempDir.z < 0.0F )
 		Angle = 180.0F - Angle;
@@ -266,7 +266,7 @@ void AI_AimAtTarget( MATRIX * InvMat , VECTOR * SPos, VECTOR * TPos )
 		AimData.Flags |= AI_CONTROL_TURNLEFT;
 	}
 	
-	Angle = (float) acos( TempDir.y );
+	Angle = (float) acosf( TempDir.y );
 	Angle = 90.0F - R2D( Angle );
 	Angle *= -1.0F;
 	if( Angle > 180.0F )
